% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cov.R
\name{GenCov_re_construct}
\alias{GenCov_re_construct}
\title{Function that takes a formula and data objects and creates a flexible `GenCovINLA_re` object, storing
all the information needed to run the `GenCovINLA` model, e.g. fixed and random effects.}
\usage{
GenCov_re_construct(formula, data, cov_data)
}
\arguments{
\item{formula}{A `GenCovINLA` formula object specifying a model.}

\item{data}{A data.frame with the data to use in the model}

\item{cov_data}{A named list containing the covariance matrices referenced in the formula. 
List element names should be match those reference in the formula (e.g. after any "__"). Note that
all matrices in `cov_data` must have named columns and rows corresponding to labels in `data` (e.g. before any "__")}
}
\description{
Function that takes a formula and data objects and creates a flexible `GenCovINLA_re` object, storing
all the information needed to run the `GenCovINLA` model, e.g. fixed and random effects.
}
